requirements.txt
requests >= 2.32.4
PyYAML >= 6.0.1
tqdm >= 4.66.1
retry >= 0.9.2
loguru >= 0.7.2
urllib3 >= 2.2.0
httpx >= 0.27.0
